Create new digest email for forum moderators
Dependencies: This command grabs a list of flagged posts by running a
Rake command (currently via Heroku). The list of moderators responsible
for each course is retrieved via a new endpoint for the user_api.

This commit creates a new `forums_digest_flagged` management command.
This can be used to send a daily digest with a list of flagged forum
posts to the course moderators.

This feature can be scheduled to run at regular intervals, though it is
disabled by default. It can be enabled by specifying a positive
value for settings.FLAGGED_FORUM_DIGEST_TASK_INTERVAL (default=0/off).
This value should be set to 1440 to be run daily at midnight.

The command used to grab the list of posts can be specified via
settings.COMMAND_FETCH_FLAGGED_FORUM_POSTS, e.g.,
    heroku run rake flags:flagged --app=APP_NAME_HERE
or
    bundle exec rake flags:flagged